Don't write result files for SKIPPED tests

+++ storage/ndb/test/run-test/main.cpp
@@ -1018,7 +1018,9 @@ void test_case_results(TestResult *test_result, const atrt_testcase &testcase) {
   res_dir.assfmt("result.%d", testcase.test_no);
   remove_dir(res_dir.c_str(), true);
 
-  if (testcase.m_report || test_result->result != ERR_OK) {
+  if (testcase.m_report ||
+      (test_result->result != ERR_OK &&
+        test_result->result != ERR_SKIPPED)) {
     if (rename("result", res_dir.c_str()) != 0) {
       g_logger.critical("Failed to rename %s as %s", "result", res_dir.c_str());
       remove_dir("result", true);
